Fix URL for OpenSSL source code download
The download URL has changed. I split the `OPENSSL_VERSION`
variable to be the concatentation of `OPENSSL_SERIES` (such as
`1.0.2`) and `OPENSSL_RELEASE` (such as `d`). This simplifies
constructing the proper download URL for `wget`, which also
changed slightly overall. (I made up the series / release
terminology on spot, though).